change in events api

+++ src/main/java/com/miu/waa/services/impl/EventServiceImpl.java
@@ -1,11 +1,13 @@
 package com.miu.waa.services.impl;
 
 import com.miu.waa.dto.request.EventCreateDto;
+import com.miu.waa.dto.request.EventFilterDto;
+import com.miu.waa.dto.response.EventAttandenceResponseDto;
 import com.miu.waa.dto.response.EventResponseDto;
-import com.miu.waa.entities.Event;
-import com.miu.waa.entities.EventStatus;
-import com.miu.waa.entities.User;
+import com.miu.waa.dto.response.UpcomingEventResponseDto;
+import com.miu.waa.entities.*;
 import com.miu.waa.mapper.EventDtoMapper;
+import com.miu.waa.repositories.EventAttendanceRepository;
 import com.miu.waa.repositories.EventRepository;
 import com.miu.waa.repositories.UserRepository;
 import com.miu.waa.services.EventService;
@@ -18,27 +20,16 @@ import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Optional;
 import java.util.stream.Collectors;
-
+@RequiredArgsConstructor
 @Service
 public class EventServiceImpl implements EventService {
     private final EventRepository eventRepository;
+    private final EventAttendanceRepository eventAttendanceRepository;
     private final UserRepository userRepository;
-    private final Long currentUserId;
-
-    public EventServiceImpl(EventRepository eventRepository,
-                            UserRepository userRepository) {
-        currentUserId=Long.valueOf(1);
-        this.eventRepository = eventRepository;
-        this.userRepository = userRepository;
-    }
-
-//    private Optional<Event> getEventById(Long eventId) {
-//        Optional<User> userLogin = RequestUtil.getUserLogin(request);
-//    }
 
     @Override
-    public List<EventResponseDto> findAll() {
-        return eventRepository.findAll().stream()
+    public List<EventResponseDto> findAll(EventFilterDto dto) {
+        return eventRepository.findAllEvents(dto).stream()
                 .map(EventDtoMapper.dtoMapper::eventToEventResponseDto)
                 .collect(Collectors.toList());
     }
@@ -46,10 +37,14 @@ public class EventServiceImpl implements EventService {
     @Override
     public EventResponseDto findById(Long id) {
         try{
-            Event event=eventRepository.findById(id)
-                    .orElseThrow(()->new NoSuchElementException("Event not found!!!"));
-
-            return EventDtoMapper.dtoMapper.eventToEventResponseDto(event);
+            Event event = eventRepository.findById(id)
+                    .orElseThrow(() -> new NoSuchElementException("Event not found with ID: " + id));
+            User user = RequestUtil.getUserLogin(null)
+                    .orElseThrow(() -> new NoSuchElementException("User not logged in. Please log in to continue."));
+            if (user.getRole().equals(UserRole.ADMIN) || event.getCreatedBy().getId().equals(user.getId())) {
+                return EventDtoMapper.dtoMapper.eventToEventResponseDto(event);
+            }
+            throw new NoSuchElementException("You do not have permission to access this event.");
         }
         catch(NoSuchElementException e){
             throw e;
@@ -59,7 +54,11 @@ public class EventServiceImpl implements EventService {
     @Override
     public EventResponseDto save(Event event) {
         try{
+            User user = RequestUtil.getUserLogin(null)
+                    .orElseThrow(() -> new NoSuchElementException("User not logged in. Please log in to continue."));
+
             event.setStatus(EventStatus.DRAFT);
+            event.setCreatedBy(user);
             event.setCreatedOn(LocalDateTime.now());
             event = eventRepository.save(event);
             return EventDtoMapper.dtoMapper.eventToEventResponseDto(event);
@@ -79,6 +78,12 @@ public class EventServiceImpl implements EventService {
             Event event=eventRepository.findById(id)
                     .orElseThrow(()->new NoSuchElementException("Event not found!!!"));
 
+            User user = RequestUtil.getUserLogin(null)
+                    .orElseThrow(() -> new NoSuchElementException("User not logged in. Please log in to continue."));
+
+            if (!user.getRole().equals(UserRole.ADMIN) && !event.getCreatedBy().getId().equals(user.getId())) {
+                throw new NoSuchElementException("You do not have permission to access this event.");
+            }
             if(event.getStatus()!=EventStatus.DRAFT){
                 throw new Exception("Event can not be updated!!!");
             }
@@ -93,6 +98,30 @@ public class EventServiceImpl implements EventService {
             throw e;
         }
     }
+    @Override
+    public EventResponseDto publishEvent(Long eventId) throws Exception {
+        try{
+            Event event=eventRepository.findById(eventId)
+                    .orElseThrow(()->new NoSuchElementException("Event not found!!!"));
+
+            if(event.getStatus()!=EventStatus.DRAFT){
+                throw new Exception("Event status can not be updated!!!");
+            }
+            User currentUser= RequestUtil.getUserLogin(null)
+                    .orElseThrow(()->new NoSuchElementException("User not found!!!"));
+
+            event.setStatus(EventStatus.PUBLISHED);
+            event.setApprovedOn(LocalDateTime.now());
+            event.setApprovedBy(currentUser);
+
+            event=eventRepository.save(event);
+            return EventDtoMapper.dtoMapper.eventToEventResponseDto(event);
+        }
+        catch (Exception e){
+            throw e;
+        }
+    }
+
 
     @Override
     public EventResponseDto updateStatus(Long eventId,EventStatus eventStatus) throws Exception {
@@ -100,18 +129,20 @@ public class EventServiceImpl implements EventService {
             Event event=eventRepository.findById(eventId)
                     .orElseThrow(()->new NoSuchElementException("Event not found!!!"));
 
-            if(eventStatus==EventStatus.DRAFT ||
-                (eventStatus==EventStatus.PUBLISHED && event.getStatus()!=EventStatus.DRAFT) ||
+            User user = RequestUtil.getUserLogin(null)
+                    .orElseThrow(() -> new NoSuchElementException("User not logged in. Please log in to continue."));
+
+            if (!user.getRole().equals(UserRole.ADMIN) && !event.getCreatedBy().getId().equals(user.getId())) {
+                throw new NoSuchElementException("You do not have permission to access this event.");
+            }
+
+            if(eventStatus==EventStatus.DRAFT || eventStatus==EventStatus.PUBLISHED ||
                 (eventStatus==EventStatus.CLOSED && event.getStatus()!=EventStatus.STARTED) ||
                 (eventStatus!=EventStatus.PUBLISHED && event.getStatus()!=EventStatus.PUBLISHED)
             ){
                 throw new Exception("Event status can not be updated!!!");
             }
             event.setStatus(eventStatus);
-            if(eventStatus==EventStatus.PUBLISHED){
-                event.setPublishedOn(LocalDateTime.now());
-            }
-
             event=eventRepository.save(event);
             return EventDtoMapper.dtoMapper.eventToEventResponseDto(event);
         }
@@ -125,10 +156,69 @@ public class EventServiceImpl implements EventService {
         Event event=eventRepository.findById(eventId)
                 .orElseThrow(()->new NoSuchElementException("Event not found!!!"));
 
+        User user = RequestUtil.getUserLogin(null)
+                .orElseThrow(() -> new NoSuchElementException("User not logged in. Please log in to continue."));
+
+        if (!user.getRole().equals(UserRole.ADMIN) && !event.getCreatedBy().getId().equals(user.getId())) {
+            throw new NoSuchElementException("You do not have permission to access this event.");
+        }
+
         if(event.getStatus()!=EventStatus.DRAFT){
             throw new Exception("Event can not be deleted");
         }
+
         eventRepository.delete(event);
         return EventDtoMapper.dtoMapper.eventToEventResponseDto(event);
     }
+
+    @Override
+    public List<UpcomingEventResponseDto> findAllUpcomingPublishedEvent() {
+        List<Event> publishedEvent=eventRepository.findAllUpcomingPublishedEvent();
+        return publishedEvent.stream()
+                .map(EventDtoMapper.dtoMapper::eventToUpcomingEventResponseDto)
+                .collect(Collectors.toList());
+    }
+
+    @Override
+    public void joinEvent(Long eventId) {
+        Event event=eventRepository.findById(eventId)
+                .orElseThrow(()->new NoSuchElementException("Event not found!!!"));
+
+        StringBuilder sb=new StringBuilder();
+        if(event.getStatus()!= EventStatus.STARTED){
+            sb.append("Event has not started by host");
+        }
+        else if(event.getEventDateTime().isAfter(LocalDateTime.now())){
+            sb.append("you cannot join this event");
+        }
+        if(!sb.isEmpty())
+            throw new NoSuchElementException(sb.toString());
+
+        User user= RequestUtil.getUserLogin(null)
+                .orElseThrow(()->new NoSuchElementException("User not loggedIn!!!"));
+
+        EventAttendance attendance=new EventAttendance();
+        attendance.setUser(user);
+        attendance.setEvent(event);
+        attendance.setCheckInTime(LocalDateTime.now());
+        eventAttendanceRepository.save(attendance);
+    }
+
+    @Override
+    public List<EventAttandenceResponseDto> findEventAttendance(Long eventId) {
+        Event event=eventRepository.findById(eventId)
+                .orElseThrow(()->new NoSuchElementException("Event not found!!!"));
+        User user= RequestUtil.getUserLogin(null)
+                .orElseThrow(()->new NoSuchElementException("User not loggedIn!!!"));
+
+        if (!user.getRole().equals(UserRole.ADMIN) && !event.getCreatedBy().getId().equals(user.getId())) {
+            throw new NoSuchElementException("You do not have permission to access this operation.");
+        }
+
+        List<EventAttendance> attendance=eventAttendanceRepository.findEventAttendanceByEventId(eventId);
+        return attendance.stream()
+                .map(EventDtoMapper.dtoMapper::eventAttandenceToResponseDto)
+                .collect(Collectors.toList());
+    }
+
 }
